Adjust neopixel locations, clear space in corners
The upper left corner required swapping package for the flash part to
a uson-8. The neopixels now have enough space that centering them on
the pads should provide enough clearance from the 3x8 block and the
corners

We should be able to have three corners covered by lego 2x2 plates to
hold the board inside the box.

Signed-off-by: Keith Packard <keithp@keithp.com>